refactor breadth search test
