+++ src/main/algorithms/search/algorithms/testHelpers/types.ts
@@ -1,0 +1,15 @@
+import type { GraphFormItemData } from '@/main/algorithms/search/graphForm/types'
+
+export type SearchSimulationResultData = {
+    steps: {
+        grid: {
+            itemData: GraphFormItemData,
+        }[][],
+        start: {
+            itemData: GraphFormItemData,
+        },
+        end: {
+            itemData: GraphFormItemData,
+        },
+    }[]
+}
